requirements.txt
jittor~=1.3.3.14
numpy~=1.21.5
opencv-python~=4.5.5.64
Pillow~=9.0.0